Add compose method for combinging two PyDAG objects
This commit adds a compose() method to the PyDAG object. This method
takes in another PyDAG object and 2 dictionaries on how to map the first
nodes in the other dag into the origal dag and the edge data to use
between the nodes in the map.

This is a first implementation and is not that efficient since it
requires looping over the graph structure of the other node to manually
reconstruct it in the original graph. Ideally in the future we should be
able to do this primarily as memory/vector operations and only adjusting
the graph structure on the boundaries. However, there is no support in
doing this via petgraph yet (see petgraph/petgraph#276) so for a first
pass this just iterates over the graph strucutre.

Related to #61